Revert last commit
`taskset --cpu-list` sets the list of processors for the task and not their number. Hence it causes complications as users has to identify which CPUs are available first. On the other hand `OMP_NUM_THREADS` and `MKL_NUM_THREADS` defines only the number of threads and the operating system sorts out which CPUs will be used.

The second option is preferred here as it causes less problems with multiple users.